backported to v4.16

+++ drivers/char/ipmi/ipmi_si_mem_io.c
@@ -1,4 +1,3 @@
-// SPDX-License-Identifier: GPL-2.0+
 
 #include <linux/io.h>
 #include "ipmi_si.h"
@@ -51,7 +50,7 @@ static unsigned char mem_inq(const struct si_sm_io *io, unsigned int offset)
 static void mem_outq(const struct si_sm_io *io, unsigned int offset,
 		     unsigned char b)
 {
-	writeq((u64)b << io->regshift, (io->addr)+(offset * io->regspacing));
+	writeq(b << io->regshift, (io->addr)+(offset * io->regspacing));
 }
 #endif
